src/api/requirements.txt
quart==0.19.6
openai==1.35.3
python-dotenv==1.0.1
azure-search-documents==11.4.0
asyncio==3.4.3
gunicorn
uvicorn
azure-identity==1.17.1
azure-monitor-opentelemetry==1.6.0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-httpx==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation-aiohttp-client==0.46b0
opentelemetry-instrumentation-openai==0.23.0
quart-cors==0.7.0
aiohttp==3.9.5
azure-storage-blob==12.20.0
azure.cosmos==4.7.0
tiktoken==0.7.0